XRDDE-970: Rename types.ts to openapi-types.ts for clarity
