+++ src/proxy-ui-api/frontend/package.json
@@ -3,8 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "rm -f src/types.ts && npm run generate-types && vue-cli-service serve",
-    "build": "rm -f src/types.ts && npm run generate-types && vue-cli-service build",
+    "serve": "rm -f src/openapi-types.ts && npm run generate-types && vue-cli-service serve",
+    "build": "rm -f src/openapi-types.ts && npm run generate-types && vue-cli-service build",
     "lint": "vue-cli-service lint",
     "build:report": "vue-cli-service build --report",
     "test:unit": "vue-cli-service test:unit",
