+++ src/proxy-ui-api/frontend/src/components/wizard/WizardPageCsrDetails.vue
@@ -100,7 +100,7 @@ import { ValidationProvider, ValidationObserver } from 'vee-validate';
 import HelpIcon from '@/components/ui/HelpIcon.vue';
 import LargeButton from '@/components/ui/LargeButton.vue';
 import SubViewTitle from '@/components/ui/SubViewTitle.vue';
-import { Key, Token } from '@/types';
+import { Key, Token } from '@/openapi-types';
 import { UsageTypes, CsrFormatTypes } from '@/global';
 import * as api from '@/util/api';
