+++ src/proxy-ui-api/frontend/src/store/modules/certificateSignRequest.ts
@@ -2,7 +2,7 @@
 import { ActionTree, GetterTree, Module, MutationTree } from 'vuex';
 import { RootState } from '../types';
 import { saveResponseAsFile } from '@/util/helpers';
-import { Key, CertificateAuthority, CsrSubjectFieldDescription } from '@/types';
+import { Key, CertificateAuthority, CsrSubjectFieldDescription } from '@/openapi-types';
 import * as api from '@/util/api';
 import { UsageTypes, CsrFormatTypes } from '@/global';
