+++ src/proxy-ui-api/frontend/src/store/modules/user.ts
@@ -2,7 +2,7 @@ import axiosAuth from '../../axios-auth';
 import axios from 'axios';
 import { ActionTree, GetterTree, Module, MutationTree } from 'vuex';
 import { RootState } from '../types';
-import { SecurityServer, Version } from '@/types';
+import { SecurityServer, Version } from '@/openapi-types';
 import { Tab } from '@/ui-types';
 import { mainTabs } from '@/global';
 import i18n from '@/i18n';
