+++ src/proxy-ui-api/frontend/src/util/helpers.ts
@@ -1,4 +1,4 @@
-import { Client } from '@/types';
+import { Client } from '@/openapi-types';
 
 // Filters an array of objects excluding specified object key
 export function selectedFilter(arr: any[], search: string, excluded?: string): any[] {
