+++ src/proxy-ui-api/frontend/src/views/CertificateDetails/CertificateDetails.vue
@@ -74,7 +74,7 @@
 import Vue from 'vue';
 import * as api from '@/util/api';
 import { UsageTypes, Permissions, PossibleActions } from '@/global';
-import { TokenCertificate } from '@/types';
+import { TokenCertificate } from '@/openapi-types';
 import SubViewTitle from '@/components/ui/SubViewTitle.vue';
 import CertificateInfo from '@/components/certificate/CertificateInfo.vue';
 import ConfirmDialog from '@/components/ui/ConfirmDialog.vue';
