+++ src/proxy-ui-api/frontend/src/views/InternalCertificateDetails/InternalCertificateDetails.vue
@@ -17,7 +17,7 @@
 import Vue from 'vue';
 import * as api from '@/util/api';
 import { UsageTypes, Permissions, PossibleActions } from '@/global';
-import { CertificateDetails } from '@/types';
+import { CertificateDetails } from '@/openapi-types';
 import SubViewTitle from '@/components/ui/SubViewTitle.vue';
 import SubViewFooter from '@/components/ui/SubViewFooter.vue';
 import CertificateInfo from '@/components/certificate/CertificateInfo.vue';
