+++ src/proxy-ui-api/frontend/src/views/KeyDetails/KeyDetails.vue
@@ -92,7 +92,7 @@ import Vue from 'vue';
 import * as api from '@/util/api';
 import { ValidationProvider, ValidationObserver } from 'vee-validate';
 import { UsageTypes, Permissions, PossibleActions } from '@/global';
-import { Key } from '@/types';
+import { Key } from '@/openapi-types';
 import SubViewTitle from '@/components/ui/SubViewTitle.vue';
 import ConfirmDialog from '@/components/ui/ConfirmDialog.vue';
 import LargeButton from '@/components/ui/LargeButton.vue';
