+++ src/proxy-ui-api/frontend/src/views/KeysAndCertificates/SecurityServerTlsCertificate/SecurityServerTlsCertificate.vue
@@ -74,7 +74,7 @@
 <script lang="ts">
 import Vue from 'vue';
 import { Permissions, RouteName } from '@/global';
-import { Key, CertificateDetails } from '@/types';
+import { Key, CertificateDetails } from '@/openapi-types';
 import * as api from '@/util/api';
 import LargeButton from '@/components/ui/LargeButton.vue';
 import SmallButton from '@/components/ui/SmallButton.vue';
