+++ src/proxy-ui-api/frontend/src/views/Service/AccessRightsDialog.vue
@@ -135,7 +135,7 @@ import Vue from 'vue';
 import * as api from '@/util/api';
 import { mapGetters } from 'vuex';
 import LargeButton from '@/components/ui/LargeButton.vue';
-import {ServiceClient} from '@/types';
+import {ServiceClient} from '@/openapi-types';
 
 enum ServiceClientTypes {
   GLOBALGROUP = 'GLOBALGROUP',
