+++ src/proxy-ui-api/frontend/src/views/Service/Endpoints/Endpoint/EndpointAccessRights.vue
@@ -82,7 +82,7 @@
   import Vue from 'vue';
   import * as api from '@/util/api';
   import SubViewTitle from '@/components/ui/SubViewTitle.vue';
-  import {Endpoint, ServiceClient} from '@/types';
+  import {Endpoint, ServiceClient} from '@/openapi-types';
   import LargeButton from '@/components/ui/LargeButton.vue';
   import ConfirmDialog from '@/components/ui/ConfirmDialog.vue';
   import AccessRightsDialog from '@/views/Service/AccessRightsDialog.vue';
