+++ src/proxy-ui-api/frontend/src/views/Settings/BackupAndRestore/RestoreBackupButton.vue
@@ -21,7 +21,7 @@
 <script lang="ts">
 import Vue from 'vue';
 import { Prop } from 'vue/types/options';
-import { Backup } from '@/types';
+import { Backup } from '@/openapi-types';
 import * as api from '@/util/api';
 import SmallButton from '@/components/ui/SmallButton.vue';
 import ConfirmDialog from '@/components/ui/ConfirmDialog.vue';
