+++ src/proxy-ui-api/frontend/src/views/Settings/SystemParameters/AddTimestampingServiceDialog.vue
@@ -71,7 +71,7 @@ import * as api from '@/util/api';
 import { Permissions } from '@/global';
 import LargeButton from '@/components/ui/LargeButton.vue';
 import { Prop } from 'vue/types/options';
-import { TimestampingService } from '@/types';
+import { TimestampingService } from '@/openapi-types';
 
 export default Vue.extend({
   name: 'AddTimestampingServiceDialog',
