+++ src/proxy-ui-api/frontend/src/views/Settings/SystemParameters/SystemParameters.vue
@@ -181,7 +181,7 @@
 <script lang="ts">
 import Vue from 'vue';
 import LargeButton from '@/components/ui/LargeButton.vue';
-import { Anchor, CertificateAuthority, TimestampingService } from '@/types';
+import { Anchor, CertificateAuthority, TimestampingService } from '@/openapi-types';
 import * as api from '@/util/api';
 import { Permissions } from '@/global';
 import TimestampingServiceRow from '@/views/Settings/SystemParameters/TimestampingServiceRow.vue';
